+++ src/Magick.NET/Native/Drawables/DrawingWand.cs
@@ -41,7 +41,7 @@ internal partial class DrawingWand : IDisposable
 
         [Throws]
         [Instance(SetsInstance = false)]
-        public partial void Bezier(PointInfoCollection coordinates, uint length);
+        public partial void Bezier(PointInfoCollection coordinates, nuint length);
 
         [Throws]
         [Instance(SetsInstance = false)]
@@ -208,11 +208,11 @@ internal partial class DrawingWand : IDisposable
 
         [Throws]
         [Instance(SetsInstance = false)]
-        public partial void Polygon(PointInfoCollection coordinates, uint length);
+        public partial void Polygon(PointInfoCollection coordinates, nuint length);
 
         [Throws]
         [Instance(SetsInstance = false)]
-        public partial void Polyline(PointInfoCollection coordinates, uint length);
+        public partial void Polyline(PointInfoCollection coordinates, nuint length);
 
         [Throws]
         [Instance(SetsInstance = false)]
@@ -276,7 +276,7 @@ internal partial class DrawingWand : IDisposable
 
         [Throws]
         [Instance(SetsInstance = false)]
-        public partial void StrokeDashArray(double[] dash, int length);
+        public partial void StrokeDashArray(double[] dash, nuint length);
 
         [Throws]
         [Instance(SetsInstance = false)]
@@ -292,7 +292,7 @@ internal partial class DrawingWand : IDisposable
 
         [Throws]
         [Instance(SetsInstance = false)]
-        public partial void StrokeMiterLimit(uint value);
+        public partial void StrokeMiterLimit(nuint value);
 
         [Throws]
         [Instance(SetsInstance = false)]
